+++ src/Windowing/Silk.NET.SDL/Structs/PfnReallocFunc.gen.cs
@@ -19,10 +19,10 @@ namespace Silk.NET.SDL
     public unsafe readonly struct PfnReallocFunc : IDisposable
     {
         private readonly void* _handle;
-        public delegate* unmanaged[Cdecl]<void*, uint, void*> Handle => (delegate* unmanaged[Cdecl]<void*, uint, void*>) _handle;
+        public delegate* unmanaged[Cdecl]<void*, nuint, void*> Handle => (delegate* unmanaged[Cdecl]<void*, nuint, void*>) _handle;
         public PfnReallocFunc
         (
-            delegate* unmanaged[Cdecl]<void*, uint, void*> ptr
+            delegate* unmanaged[Cdecl]<void*, nuint, void*> ptr
         ) => _handle = ptr;
 
         public PfnReallocFunc
@@ -35,7 +35,7 @@ namespace Silk.NET.SDL
 
         public static implicit operator nint(PfnReallocFunc pfn) => (nint) pfn.Handle;
         public static explicit operator PfnReallocFunc(nint pfn)
-            => new PfnReallocFunc((delegate* unmanaged[Cdecl]<void*, uint, void*>) pfn);
+            => new PfnReallocFunc((delegate* unmanaged[Cdecl]<void*, nuint, void*>) pfn);
 
         public static implicit operator PfnReallocFunc(ReallocFunc proc)
             => new PfnReallocFunc(proc);
@@ -43,10 +43,10 @@ namespace Silk.NET.SDL
         public static explicit operator ReallocFunc(PfnReallocFunc pfn)
             => SilkMarshal.PtrToDelegate<ReallocFunc>(pfn);
 
-        public static implicit operator delegate* unmanaged[Cdecl]<void*, uint, void*>(PfnReallocFunc pfn) => pfn.Handle;
-        public static implicit operator PfnReallocFunc(delegate* unmanaged[Cdecl]<void*, uint, void*> ptr) => new PfnReallocFunc(ptr);
+        public static implicit operator delegate* unmanaged[Cdecl]<void*, nuint, void*>(PfnReallocFunc pfn) => pfn.Handle;
+        public static implicit operator PfnReallocFunc(delegate* unmanaged[Cdecl]<void*, nuint, void*> ptr) => new PfnReallocFunc(ptr);
     }
 
-    public unsafe delegate void* ReallocFunc(void* arg0, uint arg1);
+    public unsafe delegate void* ReallocFunc(void* arg0, nuint arg1);
 }
